reimplementing circle X convex polygon collision

+++ src/com/simple/engine/BoundingArea.java
@@ -13,7 +13,7 @@ public abstract class BoundingArea {
 	
 	public abstract void move(Vector offset);
 	
-	public abstract BoundingArea spin(float angle);
+	public abstract BoundingArea spin(double angle);
 	
 	public String getTag() {
 		return tag;
