+++ src/com/simple/engine/Coordinate.java
@@ -2,11 +2,11 @@ package com.simple.engine;
 
 public class Coordinate {
 
-	private int x;
+	private double x;
 	
-	private int y;
+	private double y;
 
-	public Coordinate(int x, int y) {
+	public Coordinate(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
@@ -16,19 +16,19 @@ public class Coordinate {
 		this.y += vector.getY();
 	}
 	
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
